pyproject.toml
[tool.poetry]
name = "network-live"
version = "0.6.0"
description = "Collects all cells for all technologies for the Kcell radio network"
authors = ["Ramil Albakov <ramil.albakov2@gmail.com>"]
readme = "README.md"
packages = [{include = "network_live"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
enm-client-scripting = {path = "../local_packages/enm_client_scripting-1.22.2-py2.py3-none-any.whl"}
cx-Oracle = "^8.3.0"
deepmerge = "^1.1.0"
paramiko = "^3.1.0"
defusedxml = "^0.7.1"
anpusr-mail = {path = "../local_packages/anpusr_mail-0.1.0-py3-none-any.whl"}

[tool.poetry.scripts]
update-enm1-lte = "network_live.scripts.update_enm:update_enm1_lte"
update-enm2-lte = "network_live.scripts.update_enm:update_enm2_lte"
update-enm1-nr = "network_live.scripts.update_enm:update_enm1_nr"
update-enm2-nr = "network_live.scripts.update_enm:update_enm2_nr"
update-enm1-wcdma = "network_live.scripts.update_enm:update_enm1_wcdma"
update-enm2-wcdma = "network_live.scripts.update_enm:update_enm2_wcdma"
update-enm1-gsm = "network_live.scripts.update_enm:update_enm1_gsm"
update-enm2-gsm = "network_live.scripts.update_enm:update_enm2_gsm"
update-oss-wcdma = "network_live.scripts.update_oss:update_wcdma"
update-oss-gsm = "network_live.scripts.update_oss:update_gsm"
update-tele2-lte = "network_live.scripts.update_tele2:update_lte"
update-tele2-wcdma = "network_live.scripts.update_tele2:update_wcdma"
update-tele2-gsm = "network_live.scripts.update_tele2:update_gsm"
update-beeline-huawei-lte = "network_live.scripts.update_beeline:update_huawei_lte"
update-beeline-huawei-wcdma = "network_live.scripts.update_beeline:update_huawei_wcdma"
update-beeline-nokia-lte = "network_live.scripts.update_beeline:update_nokia_lte"
update-beeline-nokia-wcdma = "network_live.scripts.update_beeline:update_nokia_wcdma"
update-beeline-nokia-gsm = "network_live.scripts.update_beeline:update_nokia_gsm"
update-zte-wcdma = "network_live.scripts.update_zte:update_wcdma"
update-zte-gsm = "network_live.scripts.update_zte:update_gsm"
update-network-live = "network_live.scripts.update_nl:main"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.17.0"
pytest = "^7.3.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
